+++ sdk/src/tx/baseTxSender.ts
@@ -76,7 +76,12 @@ export abstract class BaseTxSender implements TxSender {
 			opts = this.opts;
 		}
 
-		const signedTx = await this.prepareTx(tx, additionalSigners, opts, preSigned);
+		const signedTx = await this.prepareTx(
+			tx,
+			additionalSigners,
+			opts,
+			preSigned
+		);
 
 		if (extraConfirmationOptions?.onSignedCb) {
 			extraConfirmationOptions.onSignedCb();
@@ -91,7 +96,6 @@ export abstract class BaseTxSender implements TxSender {
 		opts: ConfirmOptions,
 		preSigned?: boolean
 	): Promise<Transaction> {
-
 		if (preSigned) {
 			return tx;
 		}